+++ Models/Match3D.cpp
@@ -5,7 +5,7 @@ Match3D::Match3D(void)
 {
 }
 
-Match3D::Match3D(cv::Point3d queryPair, cv::Point3d trainPair)
+Match3D::Match3D(cv::Point3f queryPair, cv::Point3f trainPair)
 {
 	Match3D::queryPair = queryPair;
 	Match3D::trainPair = trainPair;
